fix(DTFS2-6910): update portal
