+++ portal/server/routes/middleware/validatePartialAuthToken/index.js
@@ -1,5 +1,5 @@
+const destroySessionAndRedirectToStart = require('../../../utils/destroy-session-and-redirect-to-start');
 const api = require('../../../api');
-const { PORTAL_URL } = require('../../../constants');
 /**
  * Validate the session contains a userToken for partial 2FA
  * @param {Object} req Request object
@@ -22,16 +22,6 @@ const validatePartialAuthToken = async (req, res, next) => {
     return destroySessionAndRedirectToStart(req, res);
   }
 };
-
-// TODO DTFS2-6770: extract
-function destroySessionAndRedirectToStart(req, res) {
-  const startPageRedirect = Boolean(process.env.START_PAGE_REDIRECT);
-  const redirectAddress = startPageRedirect ? PORTAL_URL : '/login';
-
-  req.session.destroy(() => {
-    res.redirect(redirectAddress);
-  });
-}
 
 module.exports = {
   validatePartialAuthToken,
